Create antenna-report.mdx
i added a link so people can look at already tested antennas and make a purchase decision. please double check my file as im not positive i made it so it will be a menu item under the antenna menu. this is my first time doing this so if i did it wrong i will be able to go look at the edit you made and learn from it. thanks. -zero