Cleaned up documentation, wrote some pros/cons
I edited the section to hopefully make it easier to read. I'd also propose we add a warning that the T-TEcho is very buggy currently and people should be warned to check the current status of bugs or check with the community, but I'd leave this up to some else's discretion.