Create GPS Module
I will like to add more instructions for people to hook up their modules to for specific boards. This is an example of one which I tried to be as detailed as possible for newer members, this will keep the same question from being asked on the #help channel. I have another for CardKB and buzzer modules.