Add Periphers for Heltec with GPS Module Details (#1014)
* Create GPS Module

I will like to add more instructions for people to hook up their modules to for specific boards. This is an example of one which I tried to be as detailed as possible for newer members, this will keep the same question from being asked on the #help channel. I have another for CardKB and buzzer modules.

* Update GPS Module

I've adjusted the tone to be more i informative. Would you be able to check if the it's possible to attach the picture or should I just remove it, also not sure how the formating will look.

* Add files via upload

* Update and rename GPS Module to GPS Module.mdx

Is this what you meant by file type?

* Update and rename GPS Module.mdx to peripherals.mdx

Updated the name and also copied the same formatting as the RAK peripherals. I did add the ones that I am planning of adding as tabs such as the CardKB and Buzzer options.

* Update and rename docs/hardware/devices/heltec/ESP32V3/peripherals.mdx to docs/hardware/devices/heltec/peripherals.mdx

Updated to follow the directory of the Raspberry

* Rename Heltec ESP32 V3 gps-module.webp to HeltecESP32V3-gps-module.webp

* Update peripherals.mdx

* Update peripherals.mdx

I corrected a missing bracket for the image

* Update peripherals.mdx

* Update peripherals.mdx

Corrected the levels for the headers and removed the extra m

---------

Co-authored-by: rcarteraz <robert.l.carter2@gmail.com>
Co-authored-by: Tommy Ekstrand <tekstrand@users.noreply.github.com>