Add description of how Presets map to meshes (#1767)
* Update lora.mdx

Added note that for Transmit Power the Apple apps do not show 0 as an option and that the user should pick the largest value.

* Update index.mdx

Added "what is a mesh" to clarify how radio settings determine when nodes can communicate. This section came out of confusion found on the Discord Doc channel.

* Update docs/configuration/radio/lora.mdx

According to an earlier PR this is a bug in the Apple app. I've opened an Issue there.

Co-authored-by: pdxlocations <117498748+pdxlocations@users.noreply.github.com>

* Update docs/about/overview/index.mdx

Excellent additions and changes.

Co-authored-by: rcarteraz <robert.l.carter2@gmail.com>

* Update index.mdx

added a "may" to the retransmit in response to a comment.

---------

Co-authored-by: pdxlocations <117498748+pdxlocations@users.noreply.github.com>
Co-authored-by: rcarteraz <robert.l.carter2@gmail.com>