Add Greater Boston Mesh discord (#1883)
We have created a new Discord server to support discussion and coordination around the Meshtastic net in the greater Boston area. This is intended to serve an overlapping purpose with the Boston Meshnet Facebook group (the primary online community space for the area currently; linked from the Boston Meshnet Github page on line 303) and specifically reach folks who do not use Facebook.

Co-authored-by: rcarteraz <robert.l.carter2@gmail.com>